requirements.txt
google-cloud-aiplatform==1.25.0
google-cloud-storage==2.8.0
kfp==2.0.1
google-cloud-pipeline-components==2.0.0
transformers==4.30.0
torch==2.0.0
datasets==2.12.0
pyyaml==6.0